Factor out common parts
