+++ internal/reconciler/tracepipeline/status_test.go
@@ -363,7 +363,7 @@ func TestUpdateStatus(t *testing.T) {
 			{
 				name: "healthy",
 				probe: prober.OTelPipelineProbeResult{
-					Healthy: true,
+					PipelineProbeResult: prober.PipelineProbeResult{Healthy: true},
 				},
 				expectedStatus: metav1.ConditionTrue,
 				expectedReason: conditions.ReasonFlowHealthy,
@@ -396,7 +396,7 @@ func TestUpdateStatus(t *testing.T) {
 			{
 				name: "some data dropped",
 				probe: prober.OTelPipelineProbeResult{
-					SomeDataDropped: true,
+					PipelineProbeResult: prober.PipelineProbeResult{SomeDataDropped: true},
 				},
 				expectedStatus: metav1.ConditionFalse,
 				expectedReason: conditions.ReasonSomeDataDropped,
@@ -404,8 +404,8 @@ func TestUpdateStatus(t *testing.T) {
 			{
 				name: "some data dropped shadows other problems",
 				probe: prober.OTelPipelineProbeResult{
-					SomeDataDropped: true,
-					Throttling:      true,
+					PipelineProbeResult: prober.PipelineProbeResult{SomeDataDropped: true},
+					Throttling:          true,
 				},
 				expectedStatus: metav1.ConditionFalse,
 				expectedReason: conditions.ReasonSomeDataDropped,
@@ -413,7 +413,7 @@ func TestUpdateStatus(t *testing.T) {
 			{
 				name: "all data dropped",
 				probe: prober.OTelPipelineProbeResult{
-					AllDataDropped: true,
+					PipelineProbeResult: prober.PipelineProbeResult{AllDataDropped: true},
 				},
 				expectedStatus: metav1.ConditionFalse,
 				expectedReason: conditions.ReasonAllDataDropped,
@@ -421,8 +421,8 @@ func TestUpdateStatus(t *testing.T) {
 			{
 				name: "all data dropped shadows other problems",
 				probe: prober.OTelPipelineProbeResult{
-					AllDataDropped: true,
-					Throttling:     true,
+					PipelineProbeResult: prober.PipelineProbeResult{AllDataDropped: true},
+					Throttling:          true,
 				},
 				expectedStatus: metav1.ConditionFalse,
 				expectedReason: conditions.ReasonAllDataDropped,
